backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
openai>=1.30.0
python-multipart==0.0.9
pdfplumber==0.11.4
pydantic==2.8.2
python-dotenv>=1.0.0
